navigation refactoring to incorporate headers on home bottom tab pages
