+++ navigation/level-4/FeedTabBar.js
@@ -1,7 +1,10 @@
-/* UI Kitten Component for Feed Top Tab Bar */
+/* Navigator Helper Component: 
+   ---- Provides Feed Top Tab Bar Functionality
+   ---> Passed into TopTab.Navigator as a tabBar prop in FeedTabBarNavigator.js
+*/
 import React from 'react';
 import { StyleSheet, Image } from 'react-native';
-import { TabBar, Tab, Button, Icon, Divider, TabElement, Text, Layout } from '@ui-kitten/components';
+import { Tab, Button, Icon, Layout } from '@ui-kitten/components';
 import { ThemedAwesomeTabBar } from '../../components/styledComponents/AwesomeTabBar';
 import { DrawerActions } from '@react-navigation/native';
 
@@ -68,57 +71,18 @@ const styles = StyleSheet.create({
    },
    logoLayout: {
       flex: 1,
-      // justifyContent: 'center',
       alignItems: 'center', 
-      // backgroundColor: 'blue'
    },
    emptyLayout: {
       flex: 1,
-      // justifyContent: 'center',
       alignItems: 'center', 
-      // backgroundColor: 'blue'
    },
    menu: {
       marginTop: '28%'
    },
    logo: {
       width: 25,
-      // width: '15%',
       height: 33,
-      // height: '100%',
-      // height: '35%',
       marginTop: '32%'
    }
  });
-
-// const styles = StyleSheet.create ({
-//    logo: {
-//       width: 24,
-//       height: 31,
-//       // marginTop: 40,
-//       // marginTop: '10%',
-//       justifyContent: 'center',
-//    },
-//    menu: {
-//       marginTop: '20%',
-//       justifyContent: 'flex-start',
-//    },
-//    TopContainer: {
-//       height: 75,
-//       width: '100%',
-//       alignItems: 'flex-end', // bottom of container
-//       justifyContent: 'center',
-//       flexDirection: 'row'
-//    },
-//    // const styles = StyleSheet.create({
-//       container: {
-//         flex: 1,
-//         flexDirection: 'row',
-//       },
-//       layout: {
-//         flex: 1,
-//         justifyContent: 'center',
-//         alignItems: 'center',
-//       },
-//    //  });
-// });
